修改依赖库。微调fab，wine版本管理制作中
- 依赖库：
  - commons io调低到2.4 模拟exa的版本
  - 新增gson和tukaani
- fab
  - 右侧和下侧距离改为相等，长按可以隐藏按钮。
- 新功能（未完成）：
  - wine版本管理，可联网下载（kron4ek）不同版本的wine，解压或删除。

+++ app/src/main/java/com/eltechs/ed/startupActions/InstallRecipesFromAssets.java
@@ -13,6 +13,11 @@ public class InstallRecipesFromAssets<StateClass extends ExagearImageAware> exte
     @Override // com.eltechs.axs.configuration.startup.StartupAction
     public void execute() {
         File recipesGuestDir = new File(getApplicationState().getExagearImage().getPath(), GuestContainersManager.RECIPES_GUEST_DIR);
+        //改成只解压一次吧？不然每次进入安卓界面就会清空一次
+        if(recipesGuestDir.exists()){
+            sendDone();
+            return;
+        }
         try {
             SafeFileHelpers.removeDirectory(recipesGuestDir);
             ZipInstallerAssets.installIfNecessary(getAppContext(), new ZipInstallerAssets.InstallCallback() { // from class: com.eltechs.ed.startupActions.InstallRecipesFromAssets.1
